Add secrets for test and dev environments
There are no issues with having a secret key in the repo for the dev and test environments, and it is one less thing to manage.

So adding keys for both which makes Travis setup simpler and is another step to bringing the project inline with our other Ruby services.