Integrate RSpec for unit testing
Currently the project uses [minitest](https://github.com/seattlerb/minitest) as its test framework. Nothing against minitest; its the default for Rails for example. It's just that _every other_ Defra ruby based project uses RSpec, and the [new teams](https:/github.com/DEFRA/sroc-service-team) experience is only with RSpec.

So not only to allow us to maintain it but to make it easier for Defra colleagues to help out if needed we feel we need to bring in RSpec.

This change will integrate RSpec and get it up and running whilst checking we don't break any existing stuff.

We'll then convert one of the existing tests across as an example in a subsequent PR. The intention is all new tests wil be done in RSpec and existing tests will be converted

- at any point we need to make changes to them
- as and when time permits