Update rspec config and add simplecov
This is as basic as you get it. We use a pattern of putting everything rspec requires into its own file in `spec/support` so we can keep the `spec_helper.rb` and `rails_helper.rb` as consistent as possible across projects. It also makes it clearer what extra tools a particular project is bringing it.

Calling `bundle exec rspec` is working with no issues and simplecov is reporting 0% as expected. This does cause an issue though; we really need to combine the coverage of minitest and rspec rather than overwriting previous runs.