Fix simplecov and add example rspec test
To confirm things are working in rspec we have found a basic helper method we can add a test for. This didn't appear to be covered by minitest so its an added bonus.

Whilst doing this and ensuring when we run both rspec and minitest we get a combined test coverage it highlighted the previous setup was incorrect. If you actually look at the coverage report generated

- some files were missing which should have been covered
- files in the config folder were included in the metrics (don't normally cover this folder)
- all the test/*_test.rb files were being included. As they are run under 'test' it was logging 100% test coverage for each file

The the end result was even though more files were being tested, they were the wrong files to be tested and they were reporting back incorrect coverage data. This is why the simplecov reporting has changed from

- 91.58% - 217 files in total. 7260 relevant lines. 6649 lines covered and 611 lines missed

to

- 55.93% - 170 files in total. 5183 relevant lines. 2899 lines covered and 2284 lines missed