Add authorisation service
First thing of note is the namespace. IMHO api connector was too generic. I feel `ChargingModule` more accurately denotes the code in this namespace is specific to the Charging Module API and not just API's in general. Should we ever need to connect to another API then we should look to refactor some of this code out (don't tell me, we probably already do connect to API's somewhere else in the code 🙉).

The next pattern you'll see is that I don't refer to the env vars directly in the initialiser. Instead I have named methods and I memoize the values. This stems from something in Sandi Metz's "Practical Object-Oriented Design in Ruby" (BTW an awesome book everyone should read, not just rubists!). It advises that anywhere your class needs to make calls out to an external party should be isolated in its own method.

The intention is to protect your code from changes made to those external parties that are outside of your control. If this happens you only need to make the change in one place. Plus you control what that thing is called so its makes sense in your context.

Reading from env vars is less likely to change, but sticking to the pattern helps keep things consistent and as a bonus IMHO helps with readability.

The tests will probably need a call to explain. But in essence when it comes to rspec

- use contexts to break up the tests and make the behaviour more readable. A good check of this is to run your tests using `bundle exec rspec --format documentation` and confirm they make sense as english
- use doubles to replace the things we don't manage
- use stubs for the things we do
- test only the public interface of a thing