Add custom exception for ChargeModule
Adding in preparation for the other code. What you'll notice here is that I have moved the logic for how the error message should be formatted to the error. If ideally we want our custom error to have a consistent error message we should move the logic for formatting it into the error itself.

I've also limited the message to just being about the error with other details that might be of use (taking your lead on those!) What I have avoided is any layout styling or content that would suggest the message would get shown directly to a user. Error messages are for our benefit only and should never be shown to a user. Ideally the UCD folks on your team would come up with something more appropriate and meaningful in the event of our error being thrown.

Even if you have no UCD folks on your team, it is still useful to stick to a principle that error messages will never be shown outside of logs, so avoid writing them as if they will be.