Fixture json for the bill runs API
Where you just need canned text or json, fixtures are fine. IMHO I think they are preferable to adding a bunch of code to replicate the same thing. In this case it ideally would be captured from the charging module and placed in here. If you have the chance that would be great. But this is sufficient for now to mock the behaviour of the API.