Add equivalent of api_object
So in a nutshell this is the APIObject as a module. A few bits to go through

* extend ActiveSupport::Concern

This is what allows me to just stick private at the top and get on with things. You can provide the same behaviour using vanilla mixin code but the active support concern IMHO makes the code more readable. It's something we have used across our Rails based projects.

* Only get request

I highlighted this in the previous PR. Applying the principles of [YAGNI](https://en.wikipedia.org/wiki/You_aren%27t_gonna_need_it) we should only add stuff we are actually using. And making a get request to the charging module API should be sufficient to confirm we can connect.

* Simplifies response handling

The main thing that was making things complex was the need to set @result. Now this is a module and not coupled to `ServiceObject` it no longer has to worry about that. So we can remove that issue from the code.

Next is something I have to watch out for in my code. That is premature optimization. Whether its writing for performance when you don't know if you have a performance issue, or trying to handle errors to low in the stack, its something to be avoided.

In this case its the latter. If we know in this context how an error should be handled, and what the path is if one happens then we should handle it here. I would argue in this case we don't, we are too low down. So its perfectly fine to run our code on the assumption everything is great but should something happen, let the thing that is calling us deal with it. This is because it is more likely to have a better idea of what to do if an error occurs.