Add the ListBillRunsService
Now we finally add something that uses the connector and talks to the API.

A key thing that tripped me up and I know know is the pattern used for services in this project. I'm used to `call()` returning the thing I need. I now know that in this project `call()` should always return `self`. We are then expected to call `my_service_instance.foo` to get the value we are after.

With this new found knowledge it meant some changes to the call method. Essentially now our behaviour is

- always return an instance of `self`
- if the request was successful `response` will be populated and `successful?` will be true
- if the request failed for whatever reason `response` will be nil and `failed?` will be true

The error will get logged should one be raised. But whatever is using `ListBillRunsService` can decide what needs to happen in each case.