Update seed script to create test users
We currently have an issue with the seeding of user accounts. We rely on a real email account to be used because we generate an unknownable password as part of the seeding. We need a real account so we can then the password reset mechanism to actually gain access.

This means if you were to seed a fresh environment, you could never actually use the `system@example.com` user unless you were able to intercept the email (only possible locally). The current seed design also makes it hard to write automated acceptance tests as we ideally need known users to authenticate with.

So the primary purpose for this change is to start seeding known user accounts based on roles and regimes. All will use the same known password taken from an environment variable.