Update seed script to create users
After some checking I found the service has a dependency on `system@example.com`. It seems to be a user reserved for actions the system takes automatically so must be created.

Other than that this updates `db/seeds.rb` with logic that will create users based on a `users.json` file. We only need to add new users to that and `seeds.rb` will pick them up.

The current set of users ensures

- we create a billing user specific to each regime
- we create a user for each role, and add them to all regimes to allow testing across them

Wil leave this PR here. Our intent once approved and merged is to follow it up with a change that ensures seeding is only allowed when

- the environment is not production
- if production, a special env var that permits seeding has been set