Create a rubocop todo list
The remaining violations will require much more thought and result in refactoring the code. So to allow us to move forward and avoid adding any new issues whilst not being penalised for the existing ones we create a .rubocop_todo.yml.

This is a list of known issues which rubocop will automatically exclude when you run `bundle exec rubocop`. The idea is you begin to work on the issues till such time as the todo is no longer needed.