Upgrade project to use Rails 6
https://trello.com/c/PZ2Q3LIa

The current rails version is very out of date and no longer supported or maintained. Currently it's 5.1.1 (May 2017). The latest v5 version is 5.2.4.3 (May 2020). Rails v5.2 is still receiving security fixes at least. But the project would still need to be updated to it.

The latest version of Rails is 6.0.3.4 (October 2020). Only Rails v6.0 is getting bug fixes as well as security updates.

This change covers updating the project to use Rails 6 along with any other changes needed to support it.