Update to use rails 6
This meant running a `bundle update` to rebuild the `Gemfile.lock`. The latest version of mocha, specifically minitest support is referenced by `minitest` not `mini_test` hence the need for a change there.

With this update the app still builds and tests pass. But we are flooded with deprecation warnings that need to be resolved next.