Fix code coverage
With everything being updated it included bumping simplecov to its latest (actually v0.19). The problem is from v0.18 they changed the format of the output file which broke a number of CI services including SonarCloud.

So this change locks it back to ensure we are still using a format SonarCloud recognises.