Clear up confusion about app servers
Both puma and passenger are app servers for rails. But we are referencing both in our Gemfile.

Currently, we are using passenger in production as that is what web-ops has configured the environment for. Locally Rails 5 switched from webrick to puma (it actually switched to it for everything!)

This change helps clear up some of the confusion by referencing the right app server in the right group to make it clear which is used where.