Update puma to latest
Puma is actually on version 5 whilst we were pulling in v3. This updates it to the latest.

Note - we do now get this message when running unit tests `/home/vagrant/.rbenv/versions/2.7.1/lib/ruby/gems/2.7.0/gems/puma-5.0.4/lib/puma/server.rb:94: warning: deprecated Object#=~ is called on Proc; it always returns nil` but expect that to be fixed in a release soon.