Remove tzinfo-data gem
This gem supports running the app on Windows. As this is an internal app specific to the EA we don't expect anyone else to be making use of it. Plus the whole development team and our environments are all running some nix version and not windows.

Removing this gem gets rid of a bunch of warning messages from different places in the app lifecycle.