Silence puma for Capybara tests
Whenever we run `bundle exec rails test` at somepoint we see this in the test output

```
Capybara starting Puma...
* Version 5.0.4 , codename: Spoony Bard
* Min threads: 0, max threads: 4
/home/vagrant/.rbenv/versions/2.7.1/lib/ruby/gems/2.7.0/gems/puma-5.0.4/lib/puma/server.rb:94: warning: deprecated Object#=~ is called on Proc; it always returns nil
* Listening on http://127.0.0.1:37836
```

It seems to be triggered by hitting a test that requires javascript. That and the solution I actually found on an [RSpec issue](https://github.com/rspec/rspec-rails/issues/1897).

We still get the following error appear

```
/home/vagrant/.rbenv/versions/2.7.1/lib/ruby/gems/2.7.0/gems/puma-5.0.4/lib/puma/server.rb:94: warning: deprecated Object#=~ is called on Proc; it always returns nil
```

But we are hoping this will be fixed in the next version of Puma.