Fix database config
https://trello.com/c/gKnSb5XO

We've been struggling to get the application running in our AWS `DEVELOPMENT` environment since we started to make the updates for Ruby and Rails latest versions.

A lot of the confusion is coming from a mismatch between what env vars we see the code expects to find versus what is actually in the environments. It's one of the reasons we are trying to take more ownership.

This change deals with our current scenario. The `database.yml` is hard coding the name of the DB in production. But we can see previous deployments used a version where this could be overridden. The confusion is we can't see it in the commit history.

🤦😩 Urgh!

These changes resolve the issue with the config and should get us up and running finally!