Refactor how we get the current commit
We show the current Git commit in the footer of each page. It's a handy way of confirming which version of the application is running when you are trying to investigate issues.

Currently, the functionality relies on a task being run during the deployment to generate a file which contains the current commit hash. But we know from working on other projects that also provide a similar feature, Capistrano generates the same kind of file when it deploys the app.

So this change updates the project to use the same logic as services like Waste Carriers and Waste Exemptions. This also allows us to delete unnecessary functionality from the deployment job.