Remove dependence on mimemagic gem
First, scan [Dependency on mimemagic 0.3.x no longer valid](https://github.com/rails/rails/issues/41750) for context!

In summary, **rails** depends on **activestorage** which depends on **marcel** which depends on **mimemagic**. Yesterday, the author of **mimemagic** was informed it was in violation of the GPL license. So, they chose to

- yank all versions up to and including 0.3.5 from rubygems
- release a new version under the GPL
- archive the project for all further changes/issues/pr's etc

Via the dependency chain Rails relies on **mimemagic 0.3.5** so new installs of Rails are now broken. It also can't just pull in the new version due to the clash with the GPL license.

Efforts are afoot to resolve the issue.

- A [fork of mimemagic](https://github.com/jellybob/mimemagic) which relies on the GPL source data being already installed and read at run time
- Updating **marcel** to use [MiniMime](https://github.com/rails/marcel/pull/25) and [ruby-magic](https://github.com/rails/marcel/pull/26) instead of **mimemagic**

Till the community has settled on a solution though, we need a fix in the short term. Ours is based on

- What the Dept. for Education team have done in [early-careers-framework](https://github.com/DFE-Digital/early-careers-framework/pull/178)
- What Sentry are doing [in their app](https://github.com/getsentry/sentry-ruby/issues/1362)

Instead of referencing `rails` in the `Gemfile` they are referencing just the parts needed and excluding `activestorage` and anything with a dependency on it.

This change does exactly the same thing.