Update check connection rake task
With this chanmge using the service is working and we are able to connect. However, it's highlighted this and most likely the rest of the code expects the data to be returned formatted. The sroc-charging-service does that formatting at the moment so we'll need to update our service to do the same.

This way we'll have the least impact on the rest of the code base.