Migrate chages into existing CalculateCharge
The `CalculateCharge` service was a duplicate of the `CalculationService` which is what we based our new `RulesService` on. But it has a few more features and stores the results of the call differently.

So rather than take the risk to update all the places that interact with `CalculateCharge` we instead chose to update its internal based on what we had learnt building the `RulesService`. This included adding a `self.test_connection()` method we can use instead in the rake task.