Attempt to fix tests
A number of other classes seem to depend on `CalculateCharge` though they don't use `call()`. Before this change because setting `@url` was in the initializer` there was a dependence on the env var `RULES_SERVICE_URL` existing. But when we checked our CI for `CHARGE_SERVICE_URL`, the thing its replacing we could find nothing. The key difference was no attempt is made to access it unless something used `call()`.

So, we have moved the instantiation of @url and therefore the dependence on our env vars to also only haappen if something uses `call()`.