Use puts instead of Rails.logger
Not actually sure where the previous log messages ended up. But after a quick test was reminded that these are not run within the rails process but outside of it. So, none of the logging was being output. Changing to using `puts()` will ensure we see it in the console and when run in an AWS environment it will be add to the whenever log.

It should also mean we start seeing output when running the tasks from Jenkins.