Remove Annual billing upload functionality (#416)
> Part of removing redis and resque dependencies

Recently we [Fixed the Upload Annual Billing Data page](https://github.com/DEFRA/sroc-tcm-admin/pull/411). But when we reported this fact to the team we were informed uploading annual billings is a defunct feature!

It was created to support the service going live but after that it became redundant. It's just no one thought to remove it.

As part of our work to remove the dependence on Resque and Redis we were already eyeing up the active job associated with processing annual uploads. So, this news means we can kill 2 birds with one stone; remove defunct code and the need to migrate an active job to being a rake task.

This removes everything related to uploading new files. This includes

- routes config
- controller handler for new
- controller methods to support new uploads
- views and partials to support uploading files
- active jobs
- the upload service

It also removes some handlers that did nothing as they were blank. We leave the functionality for reviewing the previous uploads in case access to this information is still needed.