Migrate ExportTransactionData tests to rspec
As part of [Migrate Data Export Job to rake](https://github.com/DEFRA/sroc-tcm-admin/pull/417) we ideally want to create rspec tests that just support the data export service and not the underlying things it calls.

The key service for exporting is `ExportTransactionData` which _is_ covered by tests using [minitest](https://github.com/seattlerb/minitest).

To be sure we have it properly covered before we stub it out in our other change we are taking this opportunity to migrate the existing tests to rspec. We are also taking the opportunity to rename it to make it consistent (adding `Service` to the name).