Add ExportDataFile factory and update regime
As part of digging into the `ExportTransactionData` service we've found that regimes have a `has_one` relationship with `ExportDataFile` records.

The `ExportTransactionData` depends on accessing a regime's `ExportDataFile` so this change adds the new factory and updates the existing `regime` factory to support it.

Note - Refer to [interconnected-associations](https://github.com/thoughtbot/factory_bot/blob/master/GETTING_STARTED.md#interconnected-associations) for the basis of how its setup.