Add transaction factories
The first step to being able to test what the service does when there is data is to create some factories for the `TransactionDetail` and `TransactionHeader` records.

Using the fixtures in the minitest suite the factories we've created appear to be the minimum needed for the service to complete without error.