Remove existing minitest test
The only thing minitest is doing which we are not is checking the export is using the correct columns. However, we feel this test is redundant. If it was essential it would be covered by the presenter. Plus it means maintaining a massive list of column names in 2 places; we are more likely to see errors because they have gone out of sync rather than a change caused a column to be dropped.

So, with the replacement rpsec tests we have added we feel confident to drop the minitest version.