Resolve first rubocop todo
[Memoization](https://en.wikipedia.org/wiki/Memoization) is a technique you can use to speed up your accessor methods. It caches the results of methods that do time-consuming work, work that only needs to be done once.

There is convention for doing this in ruby

```ruby
class User < ActiveRecord::Base
  def twitter_followers
    # assuming twitter_user.followers makes a network call
    @twitter_followers ||= twitter_user.followers
  end
end
```

So, the method name should match the memoized instance variable name. This was not the case in the service hence the need to fix it.

Note - Credits to https://www.justinweiss.com/articles/4-simple-memoization-patterns-in-ruby-and-one-gem/ for the example and explanation