Remove empty test files (#421)
On the face of it, you could look at the folders in the `test/` directory and think we have a number of things covered. Actually, when you start opening some of them up they are empty. Best guess is they were automatically added via the rails generator and never populated.

To avoid that confusion and to make it clearer what is and isn't covered by tests this change deletes any empty tests files.