Add support for generate via data export ctrlr.
When about to remove the existing jobs we spotted `RegimeTransactionExportJob` was being used in `DataExportController` to support a `/generate` endpoint.

We can now see that if `can_generate_export` in the `SystemConfig` record is true the TCM will display a generate link in the download page and the `/generate` endpoint will be enabled.

However, the default is `false`, and it currently does not appear to be enabled in any environment.

But just to be on the safe side, we've updated our `DataExportService` so that if supplied with a regime it will just run the export for it. We have also updated the control to use the service.

It does mean it won't run in the background anymore so there is a chance of timeouts. But the thing will still complete in the background and as we said, its not even turned on anywhere!