Add new devise shared partial
The instructions are to grab the view out of devise (either directly or using the generate function) and update as you need.

Then replace all the `<%= devise_error_messages! %>` in your existing Devise views to be `<%= render "devise/shared/error_messages", resource: resource %>`. A quick test and everything appears to still be working fine.