Remove unused methods from CsvExporter concern
The job we deleted was the only thing that uses `TransactionExportService` along with the `CsvExporter` concern.

When we dug into whether anyone uses the concern's `csv()` method we found nothing did, as well as `csv_opts()`. So, removing these methods which should mean we can also delete the unused service `TransactionExportService`.