Update Service to handle retrospective
In the case of CFD and PAS there was also a `NextCfdRetrospectiveReference` and `NextPasRetrospectiveReference` which do exactly the same as the other classes only return a slightly different format for the reference.

Aprt from a couple of character changes, the code was the same. So, to reduce duplication even further we have updated `NextReferenceService` to also handle returning a alternate format if the `:retrospective` flag is set.