Refactor out duplication in NextReferenceServices (#432)
As part of the work to drop the dependency on Resque and Redis we have identified that we need to update the `TransactionFileExporter` to no longer put exporting the files on a queue as an ActiveJob.

So, we started to dig into the code in the `TransactionFileExporter`. We spotted that it had a `next_[regime]_transaction_reference()` method for each regime. A quick comparison found they only differed in what underlying service they called.

```ruby
  def next_pas_transaction_reference(retrospective)
    result = if retrospective
               NextPasRetrospectiveReference.call(regime: regime, region: region)
             else
               NextPasReference.call(regime: regime, region: region)
             end
    if result.success?
      result.reference
    else
      ""
    end
  end

  def next_cfd_transaction_reference(retrospective)
    result = if retrospective
               NextCfdRetrospectiveReference.call(regime: regime, region: region)
             else
               NextCfdReference.call(regime: regime, region: region)
             end
    if result.success?
      result.reference
    else
      ""
    end
  end

  def next_wml_transaction_reference
    result = NextWmlReference.call(regime: regime, region: region)
    if result.success?
      result.reference
    else
      ""
    end
  end
```

But that wasn't the end of it. Once we started digging into those services we found they also were almost exact duplicates. They only differed based on the format of the reference generated. And the difference between a `Next[Regime1]Reference` service and a `Next[Regime]RetrospectiveReference` was a single character! 😱😱

So, before we touch the `TransactionFileExporter` we feel this is a situation we need to resolve. This change adds a new single `NextReferenceService` service to replace the existing ones, and will then update the `TransactionFileExporter` to use it.

** Notes

Our new service is going to be dependent on the `SequenceCounter.next_invoice_number()` method.

So we can stub it with confidence we have added some rspec test coverage for its existing behaviour. We also chucked in a test for `next_file_number()` for good measure!