Add new TransactionFileExportService
This service will be used to replace what the active job is doing.

Currently, a user will request a transaction file be created and exported via the UI. This will cause `TransactionFileExporter`

- create the `TransactionFile` record
- add a job on the resque queue to generate the export file for the new `TransactionFile` record

We're updating the app so that `TransactionFileExporter` just creates the `TransactionFile` record. Our new `TransactionFileExportService`, when called from rake, will then grab any `TransactionFile` records with a state of `initialised` (the default state) and process them.

The behaviour is essentially the same, the request from the user does not actually initiate the export, it queues it. We're just amending how that queuing is handled.