Resolve rubocop and test conflicts
Because we are not doing anything with the `initialize()` method in TransactionFileExportService rubocop was pushing us to remove it which makes sense.

However, the `ServiceObject` pattern that has been implemented by the previous team requires that it exists else it errors when `call()` is called.

So, this final change should get the tests passing, the service working and make Rubocop happy.