Migrate File Export Job to rake (#444)
> Part of removing redis and resque dependencies
>
> The service currently uses [Resque](https://github.com/resque/resque) for managing background jobs including email. We use [whenever](https://github.com/javan/whenever) to schedule adding 'jobs' to the queue, which means adding a record in [redis](https://redis.io/). Registered resque 'workers' then pick up and process the actual jobs.
>
> This would be great if we were dealing with thousands of background tasks. But we're not. We are dealing with just 3. Recently we have been caught out a number of times with issues owing to either jobs not logging, workers not being registered, and confusion about the process overall.
>
> There is just no need for all this additional complexity when we can replicate what other services have done and which this is already doing; just have **whenever** via cron run the task directly.

This change focuses on the `FileExportJob` and migrating it to be run from a [rake](https://github.com/ruby/rake) task.

We add a new `TransactionFileExportService`. This service will be used to replace what the active job is doing.

Currently, a user will request a transaction file be created and exported via the UI. This will cause `TransactionFileExporter` to

- create the `TransactionFile` record
- add a job on the Resque queue to generate the export file for the new `TransactionFile` record

We're updating the app so that `TransactionFileExporter` just creates the `TransactionFile` record. Our new `TransactionFileExportService`, when called from rake, will then grab any `TransactionFile` records with a state of `initialised` (the default state) and process them.

The behaviour is essentially the same, the request from the user does not actually initiate the export, it queues it. We're just amending how that queuing is handled.

** Note - Amend flash message in controllers

If only to be clearer and less confusing we've updated the flash message users see after requesting a transaction file be created. When the files have failed to generate recently there was confusion because users were seeing flash messages saying they had.