Fix dependabot PR's (attempt 2!) (#459)
It turns out what we thought would work in [Fix dependabot builds](https://github.com/DEFRA/sroc-tcm-admin/pull/458) simply doesn't. We think having both `pull_request` and `pull_request_target` as triggers means GitHub is just triggering `pull_request` as it sees them as duplicates.

So, with this change, we are trying option 1. That remove 'labeled' from `pull_request_target` as a trigger and add a big `if` condition that governs if the job can run.

> Credit to https://github.com/davidboike for this. See https://github.community/t/dependabot-doesnt-see-github-actions-secrets/167104/42