Add github changelog generator params file
The bit of JSON used to tell github_changelog_generator what sections to add and what labels to link them with is a reasonable amount of JSON. We don't reallt want to have to try and remember to add this each time.

So, if we add this [Params file](https://github.com/github-changelog-generator/github-changelog-generator#params-file) github_changelog_generator will automatically use it if the command is run from the root of the project.