Add custom sections to CHANGELOG (#462)
We have always used the tool [github-changelog-generator](https://github.com/github-changelog-generator/github-changelog-generator) to generate our `CHANGELOG.md`. It does this based on your PR history which is why we also try and ensure meaningful titles and descriptions are used.

We also use some non-default [labels](https://docs.github.com/en/issues/using-labels-and-milestones-to-track-work/managing-labels) for our PR's and have always wondered how we could include them in the changelog. Anything with a custom label has always gone under the catch-all **Merged PR's** section.

Today we finally found the answer for [how to tell it to add sections](https://github.com/github-changelog-generator/github-changelog-generator/wiki/Advanced-change-log-generation-examples#additional-options) based on labels you've added.

This change updates the `CHANGELOG.md` by regenerating it with this extra config which causes the new sections to appear. It should be noted no content is dropped. It is just more clearly grouped.