Remove request must be JSON and PATCH
Realised that by constraining the request to be JSON we were going to cause more work to be done at the Cypress end getting the headers right. Plus, if we made it a `get` we could trigger the endpoint and view the results from a browser.

It also makes the code a whole bunch simpler. So, we've dropped all the worry about the format and changed the request type to expect a `GET`.