Add rspec :focus tag check to CI workflow
RSpec has a handy feature which allows you to zero in on a specific test or group of tests.

> [..] allows you to limit a spec run to individual examples or groups you care about by tagging them with `:focus` metadata. When nothing is tagged with `:focus`, all examples get run.

This is great whilst working on a feature or change. But if you forget to remove the tag it effects the code coverage reported in our CI and often leads to at best a 'Doh!' but at worst some wasted head scratching time.

This change to our CI ensures any focus tags left in the rspec tests will be caught and flagged, failing the build and making it obvious why.